provide a way to identify easily each model: a human readable name taken from hash
Note that there is a lot of collisions as soon as there is few thousands
objects to identify.

However, that system ensure that, even if models are not yielded in the
same order, or if models are removed by a new constraint, the models
will keep the very same identifier.