change code arch + non-working implementation of bakasp as a service
First, asp-encoding related functions are moved in asp.py
then, poc.py is deleted, and moved into two modules:
 bakasp_backend, that implements the logic of the app
 bakasp_frontend, that links those logic with flask.app.route

As this last module supports the inclusion of an uid before all routes,
this effectively allows to implement a web page allowing one to specify
a configuration, that, if, valid, will create a new bakasp
implementation reachable using its uid.

This will enable of a framapad-like system, named AAS, where a bakasp instance with
its own configuration may be created, and will be accessible under its own
url, being 'service-domain.net/b/<uid>'.

A configuration of the AAS is possible, and that config system is yet to
be implemented (probably the code will reuse much of config.py).